ai/weekly_report/requirements.txt
pandas
numpy
matplotlib
scikit-learn
tqdm
Pillow
lightgbm==4.3.0
mlxtend==0.23.1
python-dotenv
openai==1.30.1
requests>=2.25.0
jinja2
pdfkit
